backend/requirements.txt
alembic==1.13.1
fastapi[all]==0.111.0
psycopg2==2.9.9
pydantic==2.7.1
pydantic-settings==2.2.1
PyJWT==2.8.0
python-dotenv==1.0.1
sqlalchemy==2.0.30
uvicorn==0.29.0